This is the latest sync since my hard drive crash.  Trying to rebuild all software that was lost.

+++ tests/src/python/test_qgspallabeling_base.py
@@ -1,18 +1,16 @@
 # -*- coding: utf-8 -*-
 """QGIS Unit tests for QgsPalLabeling: base suite setup
 
-From build dir: ctest -R PyQgsPalLabelingBase -V
-Set the following env variables when manually running tests:
-  PAL_SUITE to run specific tests (define in __main__)
-  PAL_VERBOSE to output individual test summary
-  PAL_CONTROL_IMAGE to trigger building of new control images
-  PAL_REPORT to open any failed image check reports in web browser
+From build dir, run: ctest -R PyQgsPalLabelingBase -V
+
+See <qgis-src-dir>/tests/testdata/labeling/README.rst for description.
 
 .. note:: This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.
 """
+
 __author__ = 'Larry Shaffer'
 __date__ = '07/09/2013'
 __copyright__ = 'Copyright 2013, The QGIS Project'
@@ -26,18 +24,18 @@ import glob
 import shutil
 import StringIO
 import tempfile
+
 from PyQt4.QtCore import *
 from PyQt4.QtGui import *
 
 from qgis.core import (
-    QGis,
     QgsCoordinateReferenceSystem,
     QgsDataSourceURI,
     QgsMapLayerRegistry,
     QgsMapRenderer,
+    QgsMapSettings,
     QgsPalLabeling,
     QgsPalLayerSettings,
-    QgsProject,
     QgsProviderRegistry,
     QgsVectorLayer,
     QgsRenderChecker
@@ -47,28 +45,43 @@ from utilities import (
     getQgisTestApp,
     TestCase,
     unittest,
-    expectedFailure,
     unitTestDataPath,
-    loadTestFont,
+    getTempfilePath,
+    renderMapToImage,
+    loadTestFonts,
+    getTestFont,
     openInBrowserTab
 )
 
+
 QGISAPP, CANVAS, IFACE, PARENT = getQgisTestApp()
-TESTFONT = loadTestFont()
+FONTSLOADED = loadTestFonts()
 
 PALREPORT = 'PAL_REPORT' in os.environ
 PALREPORTS = {}
 
 
+# noinspection PyPep8Naming,PyShadowingNames
 class TestQgsPalLabeling(TestCase):
 
     _TestDataDir = unitTestDataPath()
     _PalDataDir = os.path.join(_TestDataDir, 'labeling')
     _PalFeaturesDb = os.path.join(_PalDataDir, 'pal_features_v3.sqlite')
-    _TestFont = TESTFONT
+    _TestFont = getTestFont()  # Roman at 12 pt
+    """:type: QFont"""
     _MapRegistry = None
+    """:type: QgsMapLayerRegistry"""
     _MapRenderer = None
+    """:type: QgsMapRenderer"""
+    _MapSettings = None
+    """:type: QgsMapSettings"""
     _Canvas = None
+    """:type: QgsMapCanvas"""
+    _Pal = None
+    """:type: QgsPalLabeling"""
+    _PalEngine = None
+    """:type: QgsLabelingEngineInterface"""
+    _BaseSetup = False
 
     @classmethod
     def setUpClass(cls):
@@ -80,36 +93,48 @@ class TestQgsPalLabeling(TestCase):
 
         # verify that spatialite provider is available
         msg = '\nSpatialite provider not found, SKIPPING TEST SUITE'
+        # noinspection PyArgumentList
         res = 'spatialite' in QgsProviderRegistry.instance().providerList()
         assert res, msg
-
-        # load the FreeSansQGIS labeling test font
-        msg = '\nCould not load test font, SKIPPING TEST SUITE'
-        assert TESTFONT is not None, msg
 
         cls._TestFunction = ''
         cls._TestGroup = ''
         cls._TestGroupPrefix = ''
         cls._TestGroupAbbr = ''
+        cls._TestGroupCanvasAbbr = ''
         cls._TestImage = ''
+        cls._TestMapSettings = None
+        cls._Mismatch = 0
+        cls._Mismatches = dict()
+        cls._ColorTol = 0
+        cls._ColorTols = dict()
 
         # initialize class MapRegistry, Canvas, MapRenderer, Map and PAL
+        # noinspection PyArgumentList
         cls._MapRegistry = QgsMapLayerRegistry.instance()
-        # set color to match render test comparisons background
-        cls._Canvas.setCanvasColor(QColor(152, 219, 249))
-        cls._Map = cls._Canvas.map()
-        cls._Map.resize(QSize(600, 400))
         cls._MapRenderer = cls._Canvas.mapRenderer()
-        cls._CRS = QgsCoordinateReferenceSystem()
-        # default for labeling test data sources: WGS 84 / UTM zone 13N
-        cls._CRS.createFromSrid(32613)
-        cls._MapRenderer.setDestinationCrs(cls._CRS)
-        # use platform's native logical output dpi for QgsMapRenderer on launch
+
+        cls._MapSettings = cls.getBaseMapSettings()
+        osize = cls._MapSettings.outputSize()
+        cls._Canvas.resize(QSize(osize.width(), osize.height()))  # necessary?
+        # set color to match render test comparisons background
+        cls._Canvas.setCanvasColor(cls._MapSettings.backgroundColor())
 
         cls.setDefaultEngineSettings()
         msg = ('\nCould not initialize PAL labeling engine, '
                'SKIPPING TEST SUITE')
         assert cls._PalEngine, msg
+
+        cls._BaseSetup = True
+
+    @classmethod
+    def tearDownClass(cls):
+        """Run after all tests"""
+
+    def setUp(self):
+        """Run before each test."""
+        TestQgsPalLabeling.setDefaultEngineSettings()
+        self.lyr = self.defaultLayerSettings()
 
     @classmethod
     def setDefaultEngineSettings(cls):
@@ -119,20 +144,30 @@ class TestQgsPalLabeling(TestCase):
         cls._PalEngine = cls._MapRenderer.labelingEngine()
 
     @classmethod
-    def tearDownClass(cls):
-        """Run after all tests"""
-        cls.removeAllLayers()
-
-    @classmethod
     def removeAllLayers(cls):
         cls._MapRegistry.removeAllMapLayers()
+        cls._MapSettings.setLayers([])
+
+    @classmethod
+    def removeMapLayer(cls, layer):
+        if layer is None:
+            return
+        lyr_id = layer.id()
+        cls._MapRegistry.removeMapLayer(lyr_id)
+        ms_layers = cls._MapSettings.layers()
+        """:type: QStringList"""
+        if ms_layers.contains(lyr_id):
+            ms_layers.removeAt(ms_layers.indexOf(lyr_id))
+            cls._MapSettings.setLayers(ms_layers)
 
     @classmethod
     def getTestFont(cls):
         return QFont(cls._TestFont)
 
     @classmethod
-    def loadFeatureLayer(cls, table):
+    def loadFeatureLayer(cls, table, chk=False):
+        if chk and cls._MapRegistry.mapLayersByName(table):
+            return
         uri = QgsDataSourceURI()
         uri.setDatabase(cls._PalFeaturesDb)
         uri.setDataSource('', table, 'geometry')
@@ -140,11 +175,16 @@ class TestQgsPalLabeling(TestCase):
         # .qml should contain only style for symbology
         vlayer.loadNamedStyle(os.path.join(cls._PalDataDir,
                                            '{0}.qml'.format(table)))
+        # qDebug('render_lyr = {0}'.format(repr(vlayer)))
         cls._MapRegistry.addMapLayer(vlayer)
-        cls._MapRenderer.setLayerSet([vlayer.id()])
+        # place new layer on top of render stack
+        render_lyrs = [vlayer.id()]
+        render_lyrs.extend(cls._MapSettings.layers())
+        # qDebug('render_lyrs = {0}'.format(repr(render_lyrs)))
+        cls._MapSettings.setLayers(render_lyrs)
 
         # zoom to aoi
-        cls._MapRenderer.setExtent(cls.aoiExtent())
+        cls._MapSettings.setExtent(cls.aoiExtent())
         cls._Canvas.zoomToFullExtent()
         return vlayer
 
@@ -156,6 +196,47 @@ class TestQgsPalLabeling(TestCase):
         uri.setDataSource('', 'aoi', 'geometry')
         aoilayer = QgsVectorLayer(uri.uri(), 'aoi', 'spatialite')
         return aoilayer.extent()
+
+    @classmethod
+    def getBaseMapSettings(cls):
+        """
+        :rtype: QgsMapSettings
+        """
+        ms = QgsMapSettings()
+        crs = QgsCoordinateReferenceSystem()
+        """:type: QgsCoordinateReferenceSystem"""
+        # default for labeling test data: WGS 84 / UTM zone 13N
+        crs.createFromSrid(32613)
+        ms.setBackgroundColor(QColor(152, 219, 249))
+        ms.setOutputSize(QSize(420, 280))
+        ms.setOutputDpi(72)
+        ms.setFlag(QgsMapSettings.Antialiasing, True)
+        ms.setFlag(QgsMapSettings.UseAdvancedEffects, False)
+        ms.setFlag(QgsMapSettings.ForceVectorOutput, False)  # no caching?
+        ms.setDestinationCrs(crs)
+        ms.setCrsTransformEnabled(False)
+        ms.setMapUnits(crs.mapUnits())  # meters
+        ms.setExtent(cls.aoiExtent())
+        return ms
+
+    def cloneMapSettings(self, oms):
+        """
+        :param QgsMapSettings oms: Other QgsMapSettings
+        :rtype: QgsMapSettings
+        """
+        ms = QgsMapSettings()
+        ms.setBackgroundColor(oms.backgroundColor())
+        ms.setOutputSize(oms.outputSize())
+        ms.setOutputDpi(oms.outputDpi())
+        ms.setFlags(oms.flags())
+        ms.setDestinationCrs(oms.destinationCrs())
+        ms.setCrsTransformEnabled(oms.hasCrsTransformEnabled())
+        ms.setMapUnits(oms.mapUnits())
+        ms.setExtent(oms.extent())
+        ms.setOutputImageFormat(oms.outputImageFormat())
+
+        ms.setLayers(oms.layers())
+        return ms
 
     def configTest(self, prefix, abbr):
         """Call in setUp() function of test subclass"""
@@ -175,14 +256,14 @@ class TestQgsPalLabeling(TestCase):
         self._Test = '{0}_{1}'.format(self._TestGroupAbbr,
                                       self._TestFunction.replace('test_', ''))
 
-    def defaultSettings(self):
+    def defaultLayerSettings(self):
         lyr = QgsPalLayerSettings()
         lyr.enabled = True
-        lyr.fieldName = 'text'  # default in data sources
+        lyr.fieldName = 'text'  # default in test data sources
         font = self.getTestFont()
-        font.setPointSize(48)
+        font.setPointSize(32)
         lyr.textFont = font
-        lyr.textNamedStyle = 'Medium'
+        lyr.textNamedStyle = 'Roman'
         return lyr
 
     @staticmethod
@@ -202,39 +283,80 @@ class TestQgsPalLabeling(TestCase):
                     res[attr] = value
         return res
 
-    def saveContolImage(self, tmpimg=''):
-        if 'PAL_CONTROL_IMAGE' not in os.environ:
+    def controlImagePath(self, grpprefix=''):
+        if not grpprefix:
+            grpprefix = self._TestGroupPrefix
+        return os.path.join(self._TestDataDir, 'control_images',
+                            'expected_' + grpprefix,
+                            self._Test, self._Test + '.png')
+
+    def saveControlImage(self, tmpimg=''):
+        # don't save control images for RenderVsOtherOutput (Vs) tests, since
+        # those control images belong to a different test result
+        if ('PAL_CONTROL_IMAGE' not in os.environ
+                or 'Vs' in self._TestGroup):
             return
-        testgrpdir = 'expected_' + self._TestGroupPrefix
-        testdir = os.path.join(self._TestDataDir, 'control_images',
-                               testgrpdir, self._Test)
+        imgpath = self.controlImagePath()
+        testdir = os.path.dirname(imgpath)
         if not os.path.exists(testdir):
             os.makedirs(testdir)
-        imgbasepath = os.path.join(testdir, self._Test)
-        imgpath = imgbasepath + '.png'
+        imgbasepath = \
+            os.path.join(testdir,
+                         os.path.splitext(os.path.basename(imgpath))[0])
+        # remove any existing control images
         for f in glob.glob(imgbasepath + '.*'):
             if os.path.exists(f):
                 os.remove(f)
-        if tmpimg:
-            if os.path.exists(tmpimg):
-                shutil.copyfile(tmpimg, imgpath)
-        else:
-            self._Map.render()
-            self._Canvas.saveAsImage(imgpath)
+        qDebug('Control image for {0}.{1}'.format(self._TestGroup,
+                                                  self._TestFunction))
 
-    def renderCheck(self, mismatch=0, imgpath='', grpprefix=''):
+        if not tmpimg:
+            # TODO: this can be deprecated, when per-base-test-class rendering
+            #       in checkTest() is verified OK for all classes
+            qDebug('Rendering control to: {0}'.format(imgpath))
+            ms = self._MapSettings  # class settings
+            """:type: QgsMapSettings"""
+            settings_type = 'Class'
+            if self._TestMapSettings is not None:
+                ms = self._TestMapSettings  # per test settings
+                settings_type = 'Test'
+            qDebug('MapSettings type: {0}'.format(settings_type))
+
+            img = renderMapToImage(ms, parallel=False)
+            """:type: QImage"""
+            tmpimg = getTempfilePath('png')
+            if not img.save(tmpimg, 'png'):
+                os.unlink(tmpimg)
+                raise OSError('Control not created for: {0}'.format(imgpath))
+
+        if tmpimg and os.path.exists(tmpimg):
+            qDebug('Copying control to: {0}'.format(imgpath))
+            shutil.copyfile(tmpimg, imgpath)
+        else:
+            raise OSError('Control not copied to: {0}'.format(imgpath))
+
+    def renderCheck(self, mismatch=0, colortol=0, imgpath='', grpprefix=''):
         """Check rendered map canvas or existing image against control image
 
-        mismatch: number of pixels different from control, and still valid check
-        imgpath: existing image; if present, skips rendering canvas
-        grpprefix: compare test image/rendering against different test group
+        :mismatch: number of pixels different from control, and still valid
+        :colortol: maximum difference for each color component including alpha
+        :imgpath: existing image; if present, skips rendering canvas
+        :grpprefix: compare test image/rendering against different test group
         """
         if not grpprefix:
             grpprefix = self._TestGroupPrefix
+        ctl_path = self.controlImagePath(grpprefix)
+        if not os.path.exists(ctl_path):
+            raise OSError('Missing control image: {0}'.format(ctl_path))
         chk = QgsRenderChecker()
         chk.setControlPathPrefix('expected_' + grpprefix)
         chk.setControlName(self._Test)
-        chk.setMapRenderer(self._MapRenderer)
+        chk.setColorTolerance(colortol)
+        ms = self._MapSettings  # class settings
+        if self._TestMapSettings is not None:
+            ms = self._TestMapSettings  # per test settings
+        chk.setMapSettings(ms)
+        # noinspection PyUnusedLocal
         res = False
         if imgpath:
             res = chk.compareImages(self._Test, mismatch, str(imgpath))
@@ -245,6 +367,10 @@ class TestQgsPalLabeling(TestCase):
             PALREPORTS[testname] = str(chk.report().toLocal8Bit())
         msg = '\nRender check failed for "{0}"'.format(self._Test)
         return res, msg
+
+    def checkTest(self, **kwargs):
+        """Intended to be overridden in subclasses"""
+        pass
 
 
 class TestPALConfig(TestQgsPalLabeling):
@@ -278,7 +404,7 @@ class TestPALConfig(TestQgsPalLabeling):
 
     def test_layer_pal_activated(self):
         # Verify, via engine, that PAL labeling can be activated for layer
-        lyr = self.defaultSettings()
+        lyr = self.defaultLayerSettings()
         lyr.writeToLayer(self.layer)
         msg = '\nLayer labeling not activated, as reported by labelingEngine'
         self.assertTrue(self._PalEngine.willUseLayer(self.layer), msg)
@@ -286,7 +412,7 @@ class TestPALConfig(TestQgsPalLabeling):
     def test_write_read_settings(self):
         # Verify written PAL settings are same when read from layer
         # load and write default test settings
-        lyr1 = self.defaultSettings()
+        lyr1 = self.defaultLayerSettings()
         lyr1dict = self.settingsDict(lyr1)
         # print lyr1dict
         lyr1.writeToLayer(self.layer)
@@ -318,13 +444,17 @@ class TestPALConfig(TestQgsPalLabeling):
         self.assertFalse(pal.isShowingPartialsLabels())
 
 
-
+# noinspection PyPep8Naming,PyShadowingNames
 def runSuite(module, tests):
     """This allows for a list of test names to be selectively run.
     Also, ensures unittest verbose output comes at end, after debug output"""
     loader = unittest.defaultTestLoader
-    if 'PAL_SUITE' in os.environ and tests:
-        suite = loader.loadTestsFromNames(tests, module)
+    if 'PAL_SUITE' in os.environ:
+        if tests:
+            suite = loader.loadTestsFromNames(tests, module)
+        else:
+            raise Exception(
+                "\n\n####__ 'PAL_SUITE' set, but no tests specified __####\n")
     else:
         suite = loader.loadTestsFromModule(module)
     verb = 2 if 'PAL_VERBOSE' in os.environ else 0
